+++ src/main/java/licos/bson/element/village/character/BsonResultCharacter.java
@@ -42,7 +42,7 @@ public class BsonResultCharacter extends BsonAbstractCharacter {
     public BsonResultCharacter(ObjectId _id,
                                String $context,
                                String $id,
-                               long id,
+                               int id,
                                BsonName name,
                                String image,
                                boolean isMine,
