+++ src/main/scala/licos/json/element/village/character/JsonAbstractCharacter.scala
@@ -2,5 +2,5 @@ package licos.json.element.village.character
 
 import licos.json.element.village.{JsonElement, JsonName}
 
-abstract class JsonAbstractCharacter(`@context`: String, `@id`: String, id: Long, name: JsonName, image: String)
+abstract class JsonAbstractCharacter(`@context`: String, `@id`: String, id: Int, name: JsonName, image: String)
     extends JsonElement
